backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor==3.3.2
pymongo==4.6.1
pydantic>=2.10.0
pydantic-settings>=2.5.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.0
pymupdf>=1.26.0
python-docx==1.1.0
python-multipart==0.0.9
chromadb==0.4.22
openai==1.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
cryptography==42.0.0
numpy>=2.0.0
Pillow>=10.0.0  # Image processing for OCR and vision
pytesseract>=0.3.10  # OCR (optional - system works without it)
